BC-633 - refactor schools types
