+++ src/store/schools.ts
@@ -8,64 +8,7 @@ import {
 import { rootStore } from "./index";
 import { $axios } from "../utils/api";
 import AuthModule from "./auth";
-
-type Year = {
-	_id: string;
-	name: string;
-	endDate: string;
-	startDate: string;
-	years: {};
-	isTeamCreationByStudentsEnabled: boolean;
-	__v: number;
-};
-
-type County = {
-	antaresKey: string;
-	_id: string;
-	countyId: string;
-	name: string;
-	id: string;
-};
-
-type FederalState = {
-	abbreviation: string;
-	counties: County[];
-	logoUrl: string;
-	name: string;
-	__v: number;
-	_id: string;
-};
-
-export interface School {
-	_id: string;
-	name: string;
-	fileStorageType: string;
-	federalState: string;
-	county: County;
-	systems: string[];
-	updatedAt: string;
-	createdAt: string;
-	__v: number;
-	currentYear: string;
-	purpose: string;
-	features: {
-		rocketChat: boolean;
-		videoconference: boolean;
-		messenger: boolean;
-		studentVisibility: boolean;
-		messengerSchoolRoom: boolean;
-		messengerStudentRoomCreate: boolean;
-	};
-	enableStudentTeamCreation: boolean;
-	permissions: {};
-	inMaintenance: boolean;
-	documentBaseDir: string;
-	isExternal: boolean;
-	id: string;
-	years: {};
-	language?: string;
-	isTeamCreationByStudentsEnabled: boolean;
-}
+import { Year, FederalState, School } from "./types/schools";
 
 const SCHOOL_FEATURES: any = [
 	"rocketChat",
